Change theme colors from int32_t to uint16_t

+++ src/core/theme.h
@@ -57,9 +57,9 @@ public:
     String themePath = "";
 
     // Theme colors in RGB565 format
-    int32_t priColor = DEFAULT_PRICOLOR;
-    int32_t secColor = DEFAULT_PRICOLOR - 0x2000;
-    int32_t bgColor = 0x0000;
+    uint16_t priColor = DEFAULT_PRICOLOR;
+    uint16_t secColor = DEFAULT_PRICOLOR - 0x2000;
+    uint16_t bgColor = 0x0000;
 
     // UI Color
     void _setUiColor(uint16_t primary, uint16_t *secondary = nullptr, uint16_t *background = nullptr);
